Add slug field to profiles
Slug fields are initialized to usernames in the case of user profiles. In the future, we should ensure that usernames chosen are unique among slugs.  For example, if organizations don't have auth accounts, but have managers, you don't want a user choosing a username that's the same as some organization's slug, because you'll have URL conflicts.